requirements.txt

flask
flask-cors
python-dotenv
python-docx
PyMuPDF
groq
httpx
neo4j
tenacity
